+++ packages/article-skeleton/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.51.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.50.2...@times-components/article-skeleton@1.51.0) (2021-06-16)
+
+
+### Features
+
+* OPTA Football - Match Stats ([#2724](https://github.com/newsuk/times-components/issues/2724)) ([4120484](https://github.com/newsuk/times-components/commit/4120484ca972545e7ddfe57403ecb8516feec7eb))
+
+
+
+
+
+## [1.50.2](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.50.1...@times-components/article-skeleton@1.50.2) (2021-06-14)
+
+
+### Bug Fixes
+
+* OPTA football widgets ([#2722](https://github.com/newsuk/times-components/issues/2722)) ([68d3369](https://github.com/newsuk/times-components/commit/68d3369bca95545a3f1380d301d65edb11690b95))
+
+
+
+
+
+## [1.50.1](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.50.0...@times-components/article-skeleton@1.50.1) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article-skeleton
+
+
+
+
+
+# [1.50.0](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.49.3...@times-components/article-skeleton@1.50.0) (2021-06-10)
+
+
+### Features
+
+* **TE:** OPTA football widgets ([#2718](https://github.com/newsuk/times-components/issues/2718)) ([c84757e](https://github.com/newsuk/times-components/commit/c84757eb85b5f757fb57f49c1278baad4fa47ccd))
+
+
+
+
+
 ## [1.49.3](https://github.com/newsuk/times-components/compare/@times-components/article-skeleton@1.49.2...@times-components/article-skeleton@1.49.3) (2021-06-04)
 
 **Note:** Version bump only for package @times-components/article-skeleton