+++ packages/article/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.7.13](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.12...@times-components/article@7.7.13) (2021-06-16)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.12](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.11...@times-components/article@7.7.12) (2021-06-14)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.11](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.10...@times-components/article@7.7.11) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
+## [7.7.10](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.9...@times-components/article@7.7.10) (2021-06-10)
+
+**Note:** Version bump only for package @times-components/article
+
+
+
+
+
 ## [7.7.9](https://github.com/newsuk/times-components/compare/@times-components/article@7.7.8...@times-components/article@7.7.9) (2021-06-04)
 
 **Note:** Version bump only for package @times-components/article